+++ src/utils.c
@@ -267,3 +267,44 @@ int make_hist(int *h, int *g, int *hg, double *DxE, double *DpE, const char *fil
 
     return 0; // avisa que se cumplió la condición sobre los chi2
 }
+
+static void atomic_increment(int *ptr)
+{
+    __sync_fetch_and_add(ptr, 1);
+}
+
+void *work(void *range)
+{
+    int s = ((range_t *)range)->s;
+    int e = ((range_t *)range)->e;
+    double *x = ((range_t *)range)->xx;
+    double *p = ((range_t *)range)->pp;
+    int *h = ((range_t *)range)->hh;
+    int *g = ((range_t *)range)->gg;
+    int *hg = ((range_t *)range)->hghg;
+    unsigned int Ntandas = ((range_t *)range)->Ntandas;
+    int *steps = ((range_t *)range)->steps;
+    int BINS = ((range_t *)range)->BINS;
+    double DT = ((range_t *)range)->DT;
+    double M = ((range_t *)range)->M;
+    double alfa = ((range_t *)range)->alfa;
+    double pmin075 = ((range_t *)range)->pmin075;
+    double pmax075 = ((range_t *)range)->pmax075;
+    for (unsigned int j = 0; j < Ntandas; j++)
+    {
+        sem_wait(&iter_sem); // semaforo que señala que un hilo queda reservado para ejecución
+
+        iter_in_range(steps[j], s, e, x, p, DT, M, alfa, pmin075, pmax075); // avanza steps[j] pasos en el rango de partículas [s, e)
+
+        for (int i = s; i < e; i++)
+        {
+            int h_idx = (2.0 * x[i] + 1) * BINS + 2.5;
+            int g_idx = (p[i] / 3.0e-23 + 1) * BINS + 0.5;
+            atomic_increment(h + h_idx);                           // incrementa el casillero h_idx de h
+            atomic_increment(g + g_idx);                           // incrementa el casillero g_idx de g
+            atomic_increment(hg + (2 * BINS + 1) * h_idx + g_idx); // incrementa el casillero de hg
+        }
+        sem_post(&hist_sem); // semaforo que lo libera
+    }
+    return NULL;
+}
